Moved the Trades for Nyzul gear and Salvage +1.
Made 2 NPC Gold chests in the Commissions Agency in Whitegate,
one for each type so if more are added they can be easily added..
Added the 2 new npc's to custom npc/mobs.

Custom trade scripts for both.

Added animations for opening, the closing animation triggers when the player walks out of range of the chest via region control set in Zone.lua.

Made a custom message appear to the player displaying what chest is for what as the player approaches the chests.

removed old trades from the AF3 moogle in Port Jeuno.

removed ZoneCopy.lua as no longer needed and out of date.