Experimental changes to spoofed chat messaged to make it safer when scripters make mistakes; Attempting to make it simply do nothing when misused, rather than trigger breaks. Added an additional function that sends message to an engaged party/alliance from monster.
@Loki79
Take a look at changes made in Mars script, this should work for your mobs you want to make talk.

I made sure that if the monsters target is a pet, it still only sends to players. The other spoof types all expect the base entity to be a player so are ill suited for use by monsters as the original author on DarkStar forums only tested monster spoofs in onMobDeath, where "killer" is always a player. Prior to Mars, nobody had actually tested making monsters talk during combat let alone with pet jobs (bugs like this are why those mobs had commented out code and were not spawned with server start).

As soon as I saw message went only to the monsters current battle target I knew there would be problems which was why I said before that "I'll have to write a new function for that" - well now you have one! :)

Just remember not to overuse it, multiple spoofs in a short time or long messages will make clients DC or even crash (also have to test each message text because certain letter/symbol combinations are buggy).

To send a message to just 1 player instead of entire alliance, the other spoof type would need to be used in conjunction with an isPC() check in order to prevent mistakenly executing on a pet. The player to get that message would only be the mobs current target, and if current target was not a PlayerCharacter the isPC() check would cause it to not exec at all.