Reverted and redid my previous work on the mobskill "Accursed Armor".
I forgot that DarkStar stupidly requires some bits that are most easily provided by using an actual effect, making Loki's original plan of using a dummy effect the better idea. So I just fixed problems with that dummy effect implementation instead of scripting it out per-hit.

If need be a spike script can still be used to override the behavior/proc rate in the mobs script. For a true Curse2 effect I will need to edit the core anyway, which is best done in DarkStars trunk not Legions fork.