Added non retail hack some lever NPCs to try and make every door in Halvung able to be opened with a single script without a bunch of getPos() or static ID checks.
To do it proper I'd either need to rename NPC's to have separate scripts for each level, or check the positions, or static check the NPC IDs. Since we don't give a crap, I just had each level trigger a bunch of openDoor() for IDs around it. As a side effect a lever may run the openDoor() command on the wrong NPCs in addition to the correct ones, which should be harmless.